+++ composeApp/src/androidMain/kotlin/it/fast4x/rimusic/MainActivity.kt
@@ -201,14 +201,13 @@ import it.fast4x.rimusic.utils.transitionEffectKey
 import it.fast4x.rimusic.utils.useSystemFontKey
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 import me.knighthat.colorPalette
-import me.knighthat.innertube.Piped
+import me.knighthat.piped.Piped
 import me.knighthat.thumbnailShape
 import okhttp3.Call
 import okhttp3.Callback
