+++ extensions/innertube/src/main/kotlin/me/knighthat/piped/request/Player.kt
@@ -1,9 +1,9 @@
-package me.knighthat.innertube.request
+package me.knighthat.piped.request
 
 import io.ktor.http.HttpMethod
 import it.fast4x.innertube.utils.runCatchingNonCancellable
-import me.knighthat.innertube.Piped
-import me.knighthat.innertube.response.PlayerResponse
+import me.knighthat.piped.Piped
+import me.knighthat.piped.response.PlayerResponse
 
 suspend fun Piped.player( videoId: String ): Result<PlayerResponse?>? =
     runCatchingNonCancellable {
